REFACTOR: Máquina de estados para mudanças de tela

+++ script/include/Animations/ProceduralAnimation.h
@@ -38,7 +38,7 @@ void blinkAnimation(Color* currentColor, Color color1, Color color2, ProceduralA
 
 
 /** @brief Remove a opacidade de uma cor na duração especificada. */
-void fadeOut(Color* color, ProceduralAnimation* animation, int duration){
+void fadeOut(Color* color, ProceduralAnimation* animation, float duration){
     double timeElapsed = GetTime() - animation->initTime;
     color->a = (int)(255 - 255*(timeElapsed/duration));
     if(timeElapsed>=duration){
@@ -50,7 +50,7 @@ void fadeOut(Color* color, ProceduralAnimation* animation, int duration){
 
 
 /** @brief Eleva ao máximo a opacidade de uma cor na duração específica. */
-void fadeIn(Color* color, ProceduralAnimation* animation, int duration){
+void fadeIn(Color* color, ProceduralAnimation* animation, float duration){
     if(animation->running == false)
         animation->initTime = GetTime();
 
@@ -64,3 +64,23 @@ void fadeIn(Color* color, ProceduralAnimation* animation, int duration){
         return;
     }
 }
+
+
+// void fadeOutScreen(){
+//     static Color fadeOutColor = BLACK;
+//     fadeOutColor.a = 0;
+//     static ProceduralAnimation fadeOutAnimation = {0, false};
+//     static bool hasPreviousEnded = true;
+    
+//     if(hasPreviousEnded){
+//         fadeOutAnimation = (ProceduralAnimation){GetTime(), true};
+//         hasPreviousEnded = false;
+//     }
+
+//     if(fadeOutAnimation.running)
+//         fadeIn(&fadeOutColor, &fadeOutAnimation, 2.0f);
+//     else if(!hasPreviousEnded)
+//         hasPreviousEnded = true;
+
+//     DrawRectangle(0,0, LARGURA, ALTURA+ALTURAHUD, fadeOutColor);
+// }
